moved-components-to-CataloguingPage-folder
