+++ myCookingManager/frontend/src/components/CataloguingPage/CataloguingPage.jsx
@@ -3,8 +3,8 @@ import { styled } from "styled-components";
 import { useNavigate } from "react-router";
 import { useAuth0 } from "@auth0/auth0-react";
 
-import { CatalogueFlowContext } from "./contexts/CatalogueFlowContext";
-import useAutoScrollIntoView from "../utility_functions/hooks/useAutoScrollIntoView";
+import { CatalogueFlowContext } from "../contexts/CatalogueFlowContext";
+import useAutoScrollIntoView from "../../utility_functions/hooks/useAutoScrollIntoView";
 
 import UrlInput from "./UrlInput";
 import RecipePreview from "./RecipePreview";
