+++ myCookingManager/frontend/src/components/CataloguingPage/DialogueBox.jsx
